valid_email/requirements.txt
colorama==0.4.6
dnspython==2.6.1
emval==0.1.3
loguru==0.7.2
win32-setctime==1.1.0
